tasks: refactor get new task
with the itention of having a agnostic task buidler (without scores
which are specific to an automated task finder)
