+++ server/controllers/tasks/lib/get_new_tasks.js
@@ -3,7 +3,8 @@ const _ = __.require('builders', 'utils')
 const searchEntityDuplicatesSuggestions = require('./search_entity_duplicates_suggestions')
 const addOccurrencesToSuggestion = require('./add_occurrences_to_suggestion')
 const getAuthorWorksData = require('./get_author_works_data')
-const evaluateSuggestions = require('./evaluate_suggestions')
+const automerge = require('./automerge')
+const { getEntityNormalizedTerms } = __.require('controllers', 'entities/lib/terms_normalization')
 
 module.exports = entity => existingTasks => {
   const { uri: suspectUri } = entity
@@ -16,20 +17,42 @@ module.exports = entity => existingTasks => {
     const { labels: worksLabels } = suspectWorksData
     return Promise.all(newSuggestions.map(addOccurrencesToSuggestion(suspectWorksData)))
     .then(evaluateSuggestions(entity, worksLabels))
-    .then(filterOutExistingTasks(existingTasks))
-    .then(buildTaskObjects(suspectUri))
+    .then(filterNewTasks(existingTasks))
+    .then(buildTasksObjects(suspectUri, 'deduplicate'))
   })
 }
 
-const buildTaskObjects = suspectUri => suggestions => suggestions.map(suggestion => ({
-  type: 'deduplicate',
-  suspectUri,
-  suggestionUri: suggestion.uri,
-  lexicalScore: suggestion._score,
-  externalSourcesOccurrences: suggestion.occurrences
-}))
+const evaluateSuggestions = (suspect, workLabels) => suggestions => {
+  const suspectTerms = getEntityNormalizedTerms(suspect)
+  // Do not automerge if author name is in work title
+  // as it confuses occurences finding on WP pages
+  if (authorNameInWorkTitles(suspectTerms, workLabels)) return suggestions
+  const sourcedSuggestions = findSourced(suggestions)
+  if (sourcedSuggestions.length === 0) return suggestions
+  if (sourcedSuggestions.length > 1) return sourcedSuggestions
+  return automerge(suspect.uri, sourcedSuggestions[0])
+}
 
-const filterOutExistingTasks = existingTasks => suggestions => {
+const filterNewTasks = existingTasks => suggestions => {
   const existingTasksUris = _.map(existingTasks, 'suggestionUri')
   return suggestions.filter(suggestion => !existingTasksUris.includes(suggestion.uri))
 }
+
+const buildTasksObjects = (suspectUri, type) => suggestions => suggestions.map(suggestion => {
+  const { _score, uri: suggestionUri, occurrences } = suggestion
+  const newTaskObject = { type, suspectUri, suggestionUri }
+  if (_score) { newTaskObject.lexicalScore = _score }
+  if (occurrences) { newTaskObject.externalSourcesOccurrences = occurrences }
+  return newTaskObject
+})
+
+const authorNameInWorkTitles = (authorTerms, workLabels) => {
+  for (const authorLabel of authorTerms) {
+    for (const workLabel of workLabels) {
+      return workLabel.match(authorLabel)
+    }
+  }
+  return false
+}
+
+const findSourced = suggestions => suggestions.filter(sug => sug.occurrences.length > 0)
